+++ crates/bevy_ecs/src/hierarchy.rs
@@ -10,7 +10,8 @@
 use crate::reflect::{ReflectComponent, ReflectFromWorld};
 use crate::{
     bundle::Bundle,
-    component::{Component, HookContext},
+    component::Component,
+    component_lifecycle::HookContext,
     entity::Entity,
     relationship::{RelatedSpawner, RelatedSpawnerCommands},
     system::EntityCommands,
