+++ crates/bevy_input_focus/src/autofocus.rs
@@ -1,6 +1,6 @@
 //! Contains the [`AutoFocus`] component and related machinery.
 
-use bevy_ecs::{component::HookContext, prelude::*, world::DeferredWorld};
+use bevy_ecs::{component_lifecycle::HookContext, prelude::*, world::DeferredWorld};
 
 use crate::InputFocus;
